+++ unimport/constants.py
@@ -1,4 +1,7 @@
+import sys
+
+PY38_PLUS = sys.version_info >= (3, 8)
 DESCRIPTION = (
     "A linter, formatter for finding and removing unused import statements."
 )
-VERSION = "0.2.10"
+VERSION = "0.3.0"